+++ src/Container/DependencyResolver.php
@@ -89,13 +89,6 @@ class DependencyResolver
             $this->resolve($id, $entryPoint, $entryPoint);
         }
 
-        foreach ($this->definitions as $definition) {
-            var_dump([
-                "class" => $definition->getId(""),
-                "reference" => $definition->getReferenceCount(),
-            ]);
-        }
-
         return $this->definitions;
     }